Add RegisterOps function, ensure it is run when vision.h is included
