+++ torchvision/csrc/ops.h
@@ -1,0 +1,8 @@
+#ifndef TORCHVISION_OPS_H
+#define TORCHVISION_OPS_H
+
+namespace vision {
+int RegisterOps() noexcept;
+}
+
+#endif // TORCHVISION_OPS_H
